Seperated cardType into cardType and cardEmoji, put them into config.
Reworked how card types are stored, deck is now an index number relating
to two arrays in the config file "cardType" and "cardEmoji". /hand now
takes that position and returns the corresponding type and emoji
toghether.

The effect is the same but it'll be useful when I just want to grab the
emoji or just the name (I didn't wanna be using equality checks on long
strings when the emoji id can easily change if the bot is set up in a
different server, this is also less smelly code design).